# New escape algorithm with significantly less data dependency