--- a/include/simdjson/arm64/begin.h
+++ b/include/simdjson/arm64/begin.h
@@ -6,3 +6,5 @@
 #include "simdjson/arm64/numberparsing_defs.h"
 #include "simdjson/arm64/simd.h"
 #include "simdjson/arm64/stringparsing_defs.h"
+
+#define SIMDJSON_SKIP_BACKSLASH_SHORT_CIRCUIT 1