--- a/src/generic/stage1/amalgamated.h
+++ b/src/generic/stage1/amalgamated.h
@@ -1,11 +1,12 @@
 // Stuff other things depend on
 #include <generic/stage1/base.h>
+#include <generic/stage1/buf_block_reader.h>
+#include <generic/stage1/json_escape_scanner.h>
 #include <generic/stage1/json_string_scanner.h>
 #include <generic/stage1/utf8_lookup4_algorithm.h>
 #include <generic/stage1/json_scanner.h>
 
 // All other declarations
-#include <generic/stage1/buf_block_reader.h>
 #include <generic/stage1/find_next_document_index.h>
 #include <generic/stage1/json_minifier.h>
 #include <generic/stage1/json_structural_indexer.h>