--- a/src/generic/stage1/json_string_scanner.h
+++ b/src/generic/stage1/json_string_scanner.h
@@ -46,7 +46,7 @@ public:
 
 private:
   // Scans for escape characters
-  json_escape_scanner escape_scanner;
+  json_escape_scanner escape_scanner{};
   // Whether the last iteration was still inside a string (all 1's = true, all 0's = false).
   uint64_t prev_in_string = 0ULL;
 };