… include/simdjson/arm64/end.h
… include/simdjson/ppc64/end.h
… src/haswell.cpp
… src/icelake.cpp
… src/ppc64.cpp
… src/westmere.cpp
… tests/dom/basictests.cpp