This algorithm significantly reduces the `escaped` calculation from 9 to 7 bytes (it does not reduce overall instruction count, just dependency chain).

The algorithm is explained in detail [in the comments](https://github.com/simdjson/simdjson/blob/ca29d3fea3675f1a6f9fbc7ab660dfe7d0afb65d/src/generic/stage1/json_escape_scanner.h#L97-L124). Basically, it:

1. Shifts backslashes to get potentially-escaped bytes.
2. ORs in all odd bits to get the starting byte of odd sequences.
3. Uses subtraction to produce 1's for even backslashes and 0's for odd backslashes
4. XORs odd bits back out, flipping the odd and even backslash runs to 1010....

| Operation      | Even-sized   | Even-sized   | Odd-sized    | Odd-sized    |
|                | Even-aligned | Odd-aligned  | Even-aligned | Odd-aligned  |
|----------------|--------------|--------------|--------------|--------------|
| string         | `\\\\XXXX`   | `X\\\\XXX`   | `\\\nXXXX`   | `X\\\nXXX`   |
| backslash      | `11110000`   | `01111000`   | `11100000`   | `01110000`   |
| << 1 | `01111000`   | `00111100`   | `01110000`   | `00111000`   |
| \| odd         | `01111101`   | `01111101`   | `01110101`   | `01111101`   |
| backslash      | `11110000`   | `01111000`   | `11100000`   | `01110000`   |
| -              | `11110101`   | `00000101`   | `11100101`   | `00001101`   |
| ^ odd          | `10100000`   | `01010000`   | `10110000`   | `01011000`   |
|                | `\ \     `   | ` \ \    `   | `\ \n    `   | ` \ \n   `   |
|----------------|--------------|--------------|--------------|--------------|

I haven't tested the performance impact yet. As long as it doesn't go *down*, I'd still prefer this algorithm because it also produces a useful mask of the escape characters themselves alongside the escaped characters.

## Improvements

1. `escaped` and `next_is_escaped` can now be calculated independently, reducing data dependency.
  * `escaped` is down to 7 instructions from 9 (1 of which is parallel), so the CPU can skip ahead to the string prefix_xor sooner.
  * `next_is_escaped` is down to 7 instrucitons as well, so it'll be faster in a tight loop, too.
  * The total instruction count is the same (9 before, 9 now), just dependency chains have been reduced.
2. `escape` (a mask containing the escape characters--the backslashes themselves) is also produced, as part of the calculation for `next_is_escaped`. This isn't useful yet, but it will be when we want to do things like string unescaping in SIMD.