@lemire 2023-08-01
Pinging @geofflangdale (since I think Geoff had a similar strategy in mind though I might misremember)