@lemire 2023-08-01
For DOM processing (`parse` benchmark), we appear to have a slight regression of stage 1 (GCC 11, Ice Lake processor).

Before...

```
|- Stage 1
|    Speed        :  10.8301 ns per block ( 34.10%) -   0.1692 ns per byte -   1.9339 ns per structural -   5.9091 GB/s
|    Cycles       :  21.6774 per block    ( 34.75%) -   0.3387 per byte    -   3.8709 per structural    -    2.002 GHz est. frequency
|    Instructions :  70.6102 per block    ( 33.64%) -   1.1033 per byte    -  12.6087 per structural    -    3.257 per cycle
|    Misses       :     127 branch misses ( 23.07%) - 2 cache misses ( 10.67%) - 11179.00 cache references
```

After...
```
|- Stage 1
|    Speed        :  11.0633 ns per block ( 34.55%) -   0.1729 ns per byte -   1.9756 ns per structural -   5.7845 GB/s
|    Cycles       :  22.1234 per block    ( 35.09%) -   0.3457 per byte    -   3.9505 per structural    -    2.000 GHz est. frequency
|    Instructions :  71.4889 per block    ( 33.91%) -   1.1171 per byte    -  12.7656 per structural    -    3.231 per cycle
|    Misses       :     136 branch misses ( 22.82%) - 0 cache misses (  0.00%) - 11449.00 cache references
```

It is very small, something like 1 extra instruction per 64 bytes. So we can merge despite the regression, though it might be nice to play with the code a bit to make sure we can't convince the compiler to do a better job with the new algorithm.