@lemire 2023-08-01
We don't need to install clang-13: https://github.com/actions/runner-images/blob/main/images/linux/Ubuntu2204-Readme.md