@lemire 2023-08-01
@jkeiser What do you think of this diff?

```diff
index 08f7b22d..9ceb4c91 100644
--- a/src/generic/stage1/json_escape_scanner.h
+++ b/src/generic/stage1/json_escape_scanner.h
@@ -73,7 +73,7 @@ struct json_escape_scanner {
 private:
   static constexpr const uint64_t ODD_BITS = 0xAAAAAAAAAAAAAAAAULL;

-  uint64_t next_escaped_without_backslashes() noexcept {
+  simdjson_really_inline uint64_t next_escaped_without_backslashes() noexcept {
     uint64_t escaped = this->next_is_escaped;
     this->next_is_escaped = 0;
     return escaped;
@@ -148,4 +148,4 @@ private:
 } // namespace SIMDJSON_IMPLEMENTATION
 } // namespace simdjson
```

In my tests, it seems to make your PR just very slightly faster.

Note that the `next_escaped_without_backslashes()` call tends to dominate which is why optimizing it matters.