@jkeiser 2023-08-02
Ooh, yes! I intended all calls to be inline. Does that fix the instruction count difference?