@jkeiser 2023-08-02
I'm seeing an overall improvement on Haswell for stage1 (small but consistent).