@jkeiser 2023-08-02
I'll wait on merging until we've validated icelake isn't going down. There's absolutely no reason perf should go *down* with this PR; if CPUs are already saturated it's conceivable it might not go *up.*