@jkeiser 2023-08-02
I am the tiniest bit surprised that instruction count is actually down ... when I counted operations in the C++ code, they both came out to 9 (I treat a&~b as a single operation). But I'm guessing the compiler found something I didn't!