@jkeiser 2023-08-02
Instruction count is identical after simdjson_inline -> simdjson_really_inline. I don't actually recall whether we had simdjson_inline before; should we prefer one over the other?