@lemire 2023-08-02
@jkeiser At least in your test, the instruction count is down by 0.8% and the raw speed is up by 1% (that's because the number of instructions per cycle is very slightly up).

So this PR, at least in this case, makes things very slightly faster.