@jkeiser 2023-08-02
Yep! This is in line with my expectations. It's kinda hard to remember to add inline on every method, but I always get reminded how important it is.