@jkeiser 2023-08-02
@lemire after looking through gcc's assembly for this, I suspect the 1-instruction difference is due to the fact that the add_overflow() produces 2 instructions (actually 3, but one is probably a signal to the processor that doesn't get run):

```
        // prev_escaped = 0?
        xor     r13d, r13d
        // uint64_t odd_sequence_starts = backslash & ~even_bits & ~follows_escape;
        andn    rcx, rdx, rax
        and     rcx, rbp
        // prev_escaped = add_overflow(odd_sequence_starts, backslash, &sequences_starting_on_even_bits);
        add     rax, rcx
        setc    r13b
```