@Validark 2023-11-06
Sorry if this is the wrong forum in which to ask this, but I was wondering if it is possible to efficiently implement the escape-sequence check for when the bitstrings are in reversed order?

I've been working on the [Accelerated-Zig-Parser](https://github.com/Validark/Accelerated-Zig-Parser), and it would be really nice to be able to support big-endian machines without having to byte-reverse and then do count-**trailing**-zero queries. Specifically mips64 and powerpc64, which always (?) have a count-**leading**-zeroes instruction, it would be more efficient to just stay in big-endian land, which means that bitstrings of backslashes will be in the reverse order from what they would be on a little-endian machine.

`clz` is also one cycle faster than `ctz` on aarch64 and AMD x86_64 machines before Zen 4. Currently, there is an optimization I do on aarch64 where I speculatively bit-reverse my bitstrings in the producer code so the consumer code can use `clz` rather than `rbit`+`clz`. For me, this makes sense because I produce 4 bitstrings that need to be reversed, and I do, on average, ~10.5 "ctz" queries per 64-byte chunk, so I save about 6.5 rbit's per 64-byte chunk by doing the `rbit`s in the producer instead of the consumer. This optimization adds up to several milliseconds on my test bench with just over a million lines of code.

I was thinking a similar optimization should be applicable to x86_64 machines with 32 byte vectors, since I can reverse each of the 32-byte vectors with `vpshufb`+`vpermq`, and then use `clz` rather than `ctz` (`clz` is 1 cycle faster on AMD machines too). However, `vpermq` seems to be latency 6 on Zen 2 and 3, so we'd break the budget by using it twice. However, `vpermq` is 1 or 2 cycles on Zen 1/Zen+, so it could be worth it for those machines. 

Anyway, is there an efficient way to implement this algorithm when the bits are in the reverse order, or is it too big of a problem that the hardware-adder's carry goes in the wrong direction?

Thank you.