@geofflangdale 2023-11-06
That reverse-optimization on x86_64 is quite clever. Dealing with reversed bit strings is something that pops up in a few places (e.g. after using the affine transformations from GFNI as a bit matrix transpose). 

Unfortunately, I am not aware of any equivalent reversed-bits version of the tricks we use in simdjson with add. Fundamentally, most arithmetic operations (add, sub, mul) have a towards-MSB data flow.